requirements.txt
flask
openai
python-dotenv
beautifulsoup4
requests
apscheduler
cloudscraper==1.2.60
flask-socketio
eventlet
pyrebase4
firebase-admin
